+++ src/commonMain/kotlin/dev/inmo/saucenaoapi/models/ResultData.kt
@@ -159,3 +159,15 @@ data class ResultData(
     @SerialName("ext_urls")
     val extUrls: List<String> = emptyList()
 )
+
+val ResultData.froms: List<String>
+    get() = material ?.split(", ") ?: emptyList()
+
+val ResultData.authors: List<String>
+    get() = creator ?.split(", ") ?: emptyList()
+
+val ResultData.charactersList: List<String>
+    get() = characters ?.split(", ") ?: emptyList()
+
+val ResultData.titles: List<String>
+    get() = title ?.split(", ") ?: emptyList()